+++ kernel/c/gobj-c/src/stats_parser.c
@@ -4,7 +4,7 @@
  *          Stats parser
  *
  *          Copyright (c) 2017-2023 Niyamaka.
- *          Copyright (c) 2024, ArtGins.
+ *          Copyright (c) 2024-2026, ArtGins.
  *          All Rights Reserved.
 ***********************************************************************/
 #include <string.h>
@@ -193,11 +193,11 @@ PUBLIC json_t *build_stats(hgobj gobj, const char *stats, json_t *kw, hgobj src)
 {
     json_t *jn_data = json_object();
 
-    json_object_update_missing_new(jn_data, _build_stats(gobj, stats, json_incref(kw), src));
+    json_object_update_missing_new(jn_data, _build_stats(gobj, stats, kw_incref(kw), src));
 
     hgobj gobj_bottom = gobj_bottom_gobj(gobj);
     while(gobj_bottom) {
-        json_object_update_missing_new(jn_data, _build_stats(gobj_bottom, stats, json_incref(kw), src));
+        json_object_update_missing_new(jn_data, _build_stats(gobj_bottom, stats, kw_incref(kw), src));
 
         gobj_bottom = gobj_bottom_gobj(gobj_bottom);
     }